Add test coverage for new features and split api-responses.ts by domain
New tests for RequestDeduplicator, EsiDiagnostics, AsyncPaginationIterator,
and extended EsiError tests for isTimeout/retryable guards. Split the 1366-line
api-responses.ts into 29 domain-specific type files with barrel re-export for
backward compatibility.
