+++ src/types/assets.ts
@@ -1,0 +1,24 @@
+export interface CharacterAsset {
+  item_id: number;
+  type_id: number;
+  quantity: number;
+  location_id: number;
+  location_type: 'station' | 'solar_system' | 'other';
+  location_flag: string;
+  is_singleton: boolean;
+  is_blueprint_copy?: boolean;
+}
+
+export interface AssetLocation {
+  item_id: number;
+  position: {
+    x: number;
+    y: number;
+    z: number;
+  };
+}
+
+export interface AssetName {
+  item_id: number;
+  name: string;
+}
