+++ src/types/insurance.ts
@@ -1,0 +1,8 @@
+export interface InsurancePrice {
+  type_id: number;
+  levels: {
+    cost: number;
+    payout: number;
+    name: string;
+  }[];
+}
