+++ src/types/killmails.ts
@@ -1,0 +1,33 @@
+export interface KillmailSummary {
+  killmail_id: number;
+  killmail_hash: string;
+}
+
+export interface Killmail {
+  killmail_id: number;
+  killmail_time: string;
+  solar_system_id: number;
+  moon_id?: number;
+  war_id?: number;
+  victim: {
+    character_id?: number;
+    corporation_id?: number;
+    alliance_id?: number;
+    faction_id?: number;
+    ship_type_id: number;
+    damage_taken: number;
+    position?: { x: number; y: number; z: number };
+    items?: unknown[];
+  };
+  attackers: {
+    character_id?: number;
+    corporation_id?: number;
+    alliance_id?: number;
+    faction_id?: number;
+    ship_type_id?: number;
+    weapon_type_id?: number;
+    damage_done: number;
+    final_blow: boolean;
+    security_status: number;
+  }[];
+}
