+++ src/types/status.ts
@@ -1,0 +1,6 @@
+export interface ServerStatus {
+  players: number;
+  server_version: string;
+  start_time: string;
+  vip?: boolean;
+}
